# backend/docs/swagger.yaml
definitions:
...
  api.Response:
    properties:
      data: {}
      error:
        type: string
      language:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
  performance.CacheStats:
    properties:
      buffer_cache_hit_ratio:
        type: number
      index_cache_hit_ratio:
        type: number
    type: object
...
  performance.IndexUsageStats:
    properties:
      index_name:
        type: string
      index_size:
        type: string
      scans:
        type: integer
      schema_name:
        type: string
      table_name:
        type: string
      tuples_fetch:
        type: integer
      tuples_read:
        type: integer
      usage_level:
        type: string
    type: object
  performance.OptimizationRecommendation:
    properties:
      column_names:
        type: string
      estimated_benefit:
        type: string
      rationale:
        type: string
      sql_command:
        type: string
      table_name:
        type: string
      type:
        type: string
    type: object
...
  performance.TableStats:
    properties:
      deletes:
        type: integer
      idx_tuples_fetch:
        type: integer
      index_scans:
        type: integer
      inserts:
        type: integer
      schema_name:
        type: string
      seq_scan_percentage:
        type: number
      seq_tuples_read:
        type: integer
      sequential_scans:
        type: integer
      table_name:
        type: string
      total_size:
        type: string
      updates:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/performance/cache:
    get:
      description: Retrieves cache hit ratio statistics for buffer and index caches
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/performance.CacheStats'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Get cache hit ratio
      tags:
      - performance
...
  /api/v1/performance/dashboard:
    get:
      description: Provides a comprehensive performance dashboard with key metrics
        and alerts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Get performance dashboard
      tags:
      - performance
  /api/v1/performance/indexes:
    get:
      description: Retrieves detailed index usage statistics for performance analysis
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/performance.IndexUsageStats'
            type: array
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Get index usage statistics
      tags:
      - performance
...
  /api/v1/performance/optimize:
    post:
      description: Executes automated database optimization procedures
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Run database optimization
      tags:
      - performance
  /api/v1/performance/recommendations:
    get:
      description: Retrieves automated recommendations for database optimization
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/performance.OptimizationRecommendation'
            type: array
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Get optimization recommendations
      tags:
      - performance
...
  /api/v1/performance/tables:
    get:
      description: Retrieves detailed table statistics including sequential scans
        and size information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/performance.TableStats'
            type: array
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Get table statistics
      tags:
      - performance
...
swagger: "2.0"